requirements.txt
torch==1.13.1
pytest==7.2.1
pyyaml==6.0
matplotlib==3.6.3
wandb==0.13.10
streamlit==1.19.0